wip: move pino logger to framework
