+++ src/framework/pino-logger/app-logger.ts
@@ -1,6 +1,6 @@
 import { loggerConfig } from '@/config/logger';
+import { createLogger } from '@/framework/pino-logger/pino';
 import { type Logger } from '@/kairos/shared/logger/logger.type';
-import { createLogger } from '@/kairos/shared/logger/pino';
 
 export const appLogger: Logger = createLogger({
   level: loggerConfig.level,
