+++ src/index.ts
@@ -1,6 +1,6 @@
 import '@/bootstrap';
+import { appLogger } from '@/framework/pino-logger/app-logger';
 import { ensureRuntimeInfrastructure, executeSetup } from '@/kairos/setup/setup-composition';
-import { appLogger } from '@/kairos/shared/logger/app-logger';
 import { create } from '@koala-ts/framework';
 import { appConfig, server } from './config';
